requirements.txt
Flask~=2.2.3
google-cloud-dialogflow-cx~=1.20.1
numpy~=1.24.3
scipy~=1.10.1
opencv-python~=4.7.0.72